style(app): filled trash + pencil glyphs for the list rail
Consolidate the tab's local SVGs into icons.tsx: the outlined can stays with
the selection dock, while the rail's hover actions get the supplied solid
trash and pencil glyphs (both currentColor, so they inherit the muted
foreground and brighten on hover with the row).
