+++ apps/app/src/app/(dashboard)/_components/icons.tsx
@@ -1,0 +1,73 @@
+'use client';
+
+/**
+ * Local glyphs for the Lists tab. Every path draws with `currentColor`, so
+ * each call site controls the color through normal text-color utilities.
+ */
+
+/** Outlined lidded trash can — the destructive action in the selection dock. */
+export function TrashCanIcon({ className }: { className?: string }) {
+    return (
+        <svg className={className} viewBox="0 0 32 32" fill="none" aria-hidden="true" focusable="false">
+            <path
+                d="M6 9.65043V9.30309C6 8.00173 6.99637 6.91912 8.29472 6.8307C10.4245 6.68565 13.6158 6.50002 16 6.50002C18.3842 6.50002 21.5755 6.68565 23.7053 6.8307C25.0036 6.91912 26 8.00173 26 9.30309V9.65043C26 10.3308 25.4573 10.886 24.777 10.8984C22.7981 10.9346 18.8303 11 16 11C13.1697 11 9.20192 10.9346 7.22296 10.8984C6.54272 10.886 6 10.3308 6 9.65043Z"
+                stroke="currentColor"
+                strokeWidth={2}
+            />
+            <path
+                d="M18 6.53398C18.7479 6.53398 19.3542 6.58412 19.3542 6.3541C19.3542 5.6063 18.7479 5 18 5H14.2093C13.4614 5 12.8551 5.6063 12.8551 6.3541C12.8551 6.57531 13.2521 6.53398 14 6.53398H18Z"
+                stroke="currentColor"
+                strokeWidth={2}
+            />
+            <path
+                fillRule="evenodd"
+                clipRule="evenodd"
+                d="M16 27C14.4216 27 12.3348 26.7858 10.876 26.6081C9.82444 26.4801 8.9904 25.6813 8.80082 24.6391C8.51071 23.0442 8.09783 20.6636 7.87503 18.875C7.634 16.9401 7.4377 14.2717 7.33287 12.6945C7.28522 11.9777 7.85448 11.375 8.57285 11.375H16H23.4272C24.1456 11.375 24.7148 11.9777 24.6672 12.6945C24.5624 14.2717 24.3661 16.9401 24.125 18.875C23.9022 20.6636 23.4893 23.0442 23.1992 24.6391C23.0097 25.6813 22.1756 26.4801 21.1241 26.6081C19.6652 26.7858 17.5784 27 16 27Z"
+                stroke="currentColor"
+                strokeWidth={2}
+            />
+            <path
+                d="M13.4904 16.2201C13.4047 15.535 12.78 15.0491 12.095 15.1348C11.41 15.2204 10.9241 15.8451 11.0097 16.5301L11.6347 21.5301C11.7203 22.2152 12.345 22.7011 13.0301 22.6154C13.7151 22.5298 14.201 21.9051 14.1154 21.2201L13.4904 16.2201Z"
+                fill="currentColor"
+            />
+            <path
+                d="M19.9051 15.1348C20.5901 15.2204 21.076 15.8451 20.9904 16.5301L20.3654 21.5301C20.2797 22.2152 19.655 22.7011 18.97 22.6154C18.285 22.5298 17.7991 21.9051 17.8847 21.2201L18.5097 16.2201C18.5953 15.535 19.22 15.0491 19.9051 15.1348Z"
+                fill="currentColor"
+            />
+        </svg>
+    );
+}
+
+/** Solid lidded trash can — the rail's per-list delete affordance. */
+export function TrashCanFillIcon({ className }: { className?: string }) {
+    return (
+        <svg className={className} viewBox="0 0 32 32" fill="none" aria-hidden="true" focusable="false">
+            <path
+                d="M6 8.65043V8.30309C6 7.00173 6.99637 5.91912 8.29472 5.8307C10.4245 5.68565 13.6158 5.50002 16 5.50002C18.3842 5.50002 21.5755 5.68565 23.7053 5.8307C25.0036 5.91912 26 7.00173 26 8.30309V8.65043C26 9.33078 25.4573 9.88599 24.777 9.89843C22.7981 9.93462 18.8303 10 16 10C13.1697 10 9.20192 9.93462 7.22296 9.89843C6.54272 9.88599 6 9.33078 6 8.65043Z"
+                fill="currentColor"
+            />
+            <path
+                d="M18 6.7083C18.7479 6.7083 19.3542 6.102 19.3542 5.3541C19.3542 4.6063 18.7479 4 18 4H14.2093C13.4614 4 12.8551 4.6063 12.8551 5.3541C12.8551 6.102 13.4614 6.7083 14.2093 6.7083H18Z"
+                fill="currentColor"
+            />
+            <path
+                fillRule="evenodd"
+                clipRule="evenodd"
+                d="M16 28C14.4216 28 12.3348 27.7858 10.876 27.6081C9.82444 27.4801 8.9904 26.6813 8.80082 25.6391C8.51071 24.0442 8.09783 21.6636 7.87503 19.875C7.634 17.9401 7.4377 15.2717 7.33287 13.6945C7.28522 12.9777 7.85448 12.375 8.57285 12.375H16H23.4272C24.1456 12.375 24.7148 12.9777 24.6672 13.6945C24.5624 15.2717 24.3661 17.9401 24.125 19.875C23.9022 21.6636 23.4893 24.0442 23.1992 25.6391C23.0097 26.6813 22.1756 27.4801 21.1241 27.6081C19.6652 27.7858 17.5784 28 16 28ZM12.095 16.1347C12.78 16.049 13.4047 16.5349 13.4904 17.22L14.1154 22.22C14.201 22.905 13.7151 23.5297 13.0301 23.6153C12.345 23.701 11.7203 23.2151 11.6347 22.53L11.0097 17.53C10.9241 16.845 11.41 16.2203 12.095 16.1347ZM20.9904 17.53C21.076 16.845 20.5901 16.2203 19.9051 16.1347C19.22 16.049 18.5953 16.5349 18.5097 17.22L17.8847 22.22C17.7991 22.905 18.285 23.5297 18.97 23.6153C19.655 23.701 20.2797 23.2151 20.3654 22.53L20.9904 17.53Z"
+                fill="currentColor"
+            />
+        </svg>
+    );
+}
+
+/** Pencil over a page corner — the rail's per-list settings affordance. */
+export function PencilIcon({ className }: { className?: string }) {
+    return (
+        <svg className={className} viewBox="0 0 20 20" fill="none" aria-hidden="true" focusable="false">
+            <path
+                d="M6.66672 16.6668V17.5418C6.89836 17.5418 7.12055 17.45 7.28458 17.2864L6.66672 16.6668ZM3.33337 16.6668H2.45837C2.45837 17.1501 2.85012 17.5418 3.33337 17.5418V16.6668ZM3.33337 13.2725L2.71552 12.6529C2.5509 12.8171 2.45837 13.04 2.45837 13.2725H3.33337ZM13.22 4.64907C13.8082 4.06197 14.763 4.06172 15.3512 4.64882L16.5874 3.41016C15.3161 2.14127 13.2554 2.14127 11.984 3.41016L13.22 4.64907ZM15.3512 4.64882C15.9385 5.23496 15.9385 6.18457 15.3512 6.77071L16.5872 8.00962C17.8596 6.73978 17.8598 4.68 16.5874 3.41016L15.3512 4.64882ZM15.3512 6.77071L6.04886 16.0473L7.28458 17.2864L16.5872 8.00962L15.3512 6.77071ZM6.66672 15.7918H3.33337V17.5418H6.66672V15.7918ZM11.984 3.41016L2.71552 12.6529L3.95123 13.8921L13.22 4.64907L11.984 3.41016ZM2.45837 13.2725V16.6668H4.20837V13.2725H2.45837ZM10.5555 6.07458L13.9227 9.43513L15.1589 8.19647L11.7917 4.83592L10.5555 6.07458Z"
+                fill="currentColor"
+            />
+        </svg>
+    );
+}
